// frontend/types/chart.ts
export interface ChartCandle {
  id: number
  market: string
  candle_date_time_kst: string
  opening_price: number
  high_price: number
  low_price: number
  trade_price: number
  timestamp: number
  candle_acc_trade_price: number
  candle_acc_trade_volume: number
  unit: number
}

export interface ChartParams {
  market: string
  unit: number
  limit: number
}